backend/requirements.txt
alembic==1.14.1
APScheduler==3.11.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.15
filelock==3.16.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
six==1.17.0
SQLAlchemy==2.0.37
tokenizers==0.13.3
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websocket-server==0.6.4
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
